added authentication and registration
